Move tests from index to helpers.js file
