+++ client/src/common/utils/helpers.js
@@ -1,18 +1,5 @@
 import { AbortPromiseException } from 'common/exceptions/AbortPromiseException';
 
-export const isUrlValid = string => {
-  const pattern = new RegExp(
-    '^(https?:\\/\\/)?' +
-      '((([a-z\\d]([a-z\\d-]*[a-z\\d])*)\\.)+[a-z]{2,}|' +
-      '((\\d{1,3}\\.){3}\\d{1,3}))' +
-      '(\\:\\d+)?(\\/[-a-z\\d%_.~+]*)*' +
-      '(\\?[;&a-z\\d%_.~+=-]*)?' +
-      '(\\#[-a-z\\d_]*)?$',
-    'i'
-  );
-  return pattern.test(string);
-};
-
 export const makeAbortablePromise = promise => {
   let _reject;
