+++ package.json
@@ -47,7 +47,8 @@
     "react-router-dom": "4.3.1",
     "redux": "4.0.1",
     "redux-thunk": "2.3.0",
-    "reselect": "4.0.0"
+    "reselect": "4.0.0",
+    "validator": "10.11.0"
   },
   "devDependencies": {
     "@babel/core": "7.0.0",
